+++ Source/AudioSend.h
@@ -39,7 +39,7 @@ public:
 private:
    //IDrawableModule
    void DrawModule() override;
-   void GetModuleDimensions(int& w, int&h) override { w=86; h=38; }
+   void GetModuleDimensions(float& w, float& h) override { w=86; h=38; }
    bool Enabled() const override { return mEnabled; }
    
    bool mCrossfade;
