+++ Source/Inverter.h
@@ -36,6 +36,6 @@ public:
 private:
    //IDrawableModule
    void DrawModule() override;
-   void GetModuleDimensions(int& w, int&h) override { w=120; h=12; }
+   void GetModuleDimensions(float& w, float& h) override { w=120; h=12; }
    bool Enabled() const override { return mEnabled; }
 };
