+++ Source/KarplusStrong.h
@@ -52,7 +52,7 @@ public:
 private:
    //IDrawableModule
    void DrawModule() override;
-   void GetModuleDimensions(int& width, int& height) override { width = 250; height = 90; }
+   void GetModuleDimensions(float& width, float& height) override { width = 250; height = 90; }
    bool Enabled() const override { return mEnabled; }
 
    PolyphonyMgr mPolyMgr;
