+++ Source/KeyboardDisplay.h
@@ -38,7 +38,7 @@ public:
 private:
    //IDrawableModule
    void DrawModule() override;
-   void GetModuleDimensions(int& width, int& height) override { width = mWidth; height = mHeight; }
+   void GetModuleDimensions(float& width, float& height) override { width = mWidth; height = mHeight; }
    bool Enabled() const override { return mEnabled; }
    void OnClicked(int x, int y, bool right) override;
    bool IsResizable() const override { return true; }
