+++ Source/Kicker.h
@@ -38,7 +38,7 @@ public:
 private:
    //IDrawableModule
    void DrawModule() override;
-   void GetModuleDimensions(int& width, int& height) override { width = 90; height = 0; }
+   void GetModuleDimensions(float& width, float& height) override { width = 90; height = 0; }
    bool Enabled() const override { return mEnabled; }
    
    DrumPlayer* mDrumPlayer;
