+++ Source/LinnstrumentControl.h
@@ -82,7 +82,7 @@ private:
    //IDrawableModule
    void DrawModule() override;
    bool Enabled() const override { return true; }
-   void GetModuleDimensions(int& w, int& h) override { w=190; h=7+17*4; }
+   void GetModuleDimensions(float& w, float& h) override { w=190; h=7+17*4; }
    
    int mControllerIndex;
    DropdownList* mControllerList;
